move LBS FM to separate file
